requirements.txt
black==23.10.1
Bottleneck==1.3.5
click==8.1.7
cmdstanpy==1.1.0
contourpy==1.0.5
convertdate==2.3.2
cycler==0.11.0
deap==1.4.1
ephem==4.1.2
fonttools==4.25.0
holidays==0.29
importlib-resources==6.1.0
iniconfig==1.1.1
isort==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
lightgbm==4.1.0
LunarCalendar==0.0.9
matplotlib==3.8.0
munkres==1.1.4
mypy-extensions==1.0.0
numexpr==2.8.7
numpy==1.24.3
packaging==23.1
pandas==2.1.1
pathspec==0.11.2
patsy==0.5.3
Pillow==10.0.1
pip==23.3
platformdirs==4.0.0
pluggy==1.0.0
prophet==1.1.4
PyMeeus==0.5.11
pyparsing==3.0.9
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.0
scipy==1.11.3
seaborn==0.13.0
setuptools==68.0.0
six==1.16.0
sklearn-genetic-opt==0.10.1
statsmodels==0.14.0
threadpoolctl==2.2.0
tornado==6.3.3
tqdm==4.65.0
typing_extensions==4.8.0
tzdata==2023.3
wheel==0.41.2
